@@ -1,7 +1,7 @@
 {
   "name": "root",
   "private": true,
-  "license": "MIT",
+  "license": "AGPL",
   "workspaces": [
     "packages/*"
   ],
@@ -10,27 +10,27 @@
     "@types/express": "^4.17.8",
     "@types/mocha": "^8.0.3",
     "@types/node-fetch": "^2.5.7",
-    "@typescript-eslint/eslint-plugin": "^4.0.1",
-    "@typescript-eslint/parser": "^4.0.1",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
     "body-parser": "^1.19.0",
-    "eslint": "^7.8.1",
+    "eslint": "^7.10.0",
     "express": "^4.17.1",
     "lerna": "^3.20.2",
     "mocha": "^8.1.3",
     "rimraf": "^3.0.2",
     "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.0.2"
+    "typescript": "^4.0.3"
   },
   "dependencies": {
-    "@peculiar/asn1-csr": "^2.0.21",
+    "@peculiar/asn1-csr": "^2.0.22",
     "@peculiar/asn1-ecc": "^2.0.3",
-    "@peculiar/asn1-rsa": "^2.0.21",
+    "@peculiar/asn1-rsa": "^2.0.22",
     "@peculiar/asn1-schema": "^2.0.17",
-    "@peculiar/asn1-x509": "^2.0.21",
+    "@peculiar/asn1-x509": "^2.0.22",
     "@peculiar/webcrypto": "^1.1.3",
-    "node-fetch": "^2.6.0",
-    "pvtsutils": "^1.0.10",
+    "node-fetch": "^2.6.1",
+    "pvtsutils": "^1.0.14",
     "reflect-metadata": "^0.1.13",
     "tslib": "^2.0.1",
     "tsyringe": "^4.2.0"
