@@ -1,7 +1,7 @@
-import express = require("express");
-import { Express } from "express";
+import * as express from "express";
 import { container } from "tsyringe";
 import * as url from "url";
+import * as cors from "cors";
 
 import { diLogger, ILogger } from "@peculiar/acme-core";
 import { DependencyInjection as diServer, diServerOptions, IServerOptions } from "@peculiar/acme-server";
@@ -9,16 +9,25 @@ import { DependencyInjection as diServer, diServerOptions, IServerOptions } from
 import { routers } from "./routes";
 import { diControllers, Controllers } from "./controllers";
 
+export * from "./controllers";
+export * from "./routes";
+
 export type IAcmeExpressOptions = Partial<IServerOptions>;
 
 export class AcmeExpress {
 
-  public static register(app: Express, options: IAcmeExpressOptions = {}) {
+  public static register(app: express.Express, options: IAcmeExpressOptions = {}) {
     diServer.register(container, options);
 
     container.register(diControllers, Controllers);
 
     const opt = container.resolve<IServerOptions>(diServerOptions);
+
+    app.use(cors({
+      methods:"GET, POST, OPTIONS, HEAD",
+      allowedHeaders: "Content-Type, Authorization, Cache-Control, Replay-Nonce",
+      exposedHeaders: "Location, Link, Replay-Nonce"
+    }));
 
     app.use(express.json({ type: "application/jose+json" }));
     app.use(url.parse(opt.baseAddress).pathname || "/", routers);
