@@ -1,11 +1,9 @@
-import express = require("express");
+import * as express from "express";
 import { Request, Response } from 'express';
 import { container } from "tsyringe";
 import { diControllers, Controllers } from "../controllers";
 
-export const routers = express.Router({
-  strict: true
-});
+export const routers = express.Router();
 routers.get('/directory', (req: Request, res: Response) => {
   container.resolve<Controllers>(diControllers).getDirectory(req, res);
 });
