@@ -1,6 +1,6 @@
 import { AcmeError } from "./acme_error";
 import { ErrorType } from "./error_type";
-import { HttpStatusCode } from "../web";
+import { HttpStatusCode } from "../web/http_status_code";
 
 export class IncorrectResponseError extends AcmeError {
   public constructor(
