@@ -1,4 +1,4 @@
-import { cryptoProvider } from "@peculiar/acme-core";
+import { cryptoProvider } from "@peculiar/x509";
 import { INonceRepository } from "@peculiar/acme-data";
 import { Convert } from "pvtsutils";
 
