@@ -1,41 +1,39 @@
 import { Request, Response } from "express";
-import { inject, injectable } from "tsyringe";
+import { container, injectable } from "tsyringe";
 import { diAcmeController, AcmeController } from "@peculiar/acme-server";
 import { Request as AcmeRequest, Response as AcmeResponse, ContentType, MalformedError } from "@peculiar/acme-core";
-import { ParamsDictionary } from "express-serve-static-core";
 import { Key } from "@peculiar/acme-data";
 
 export const diControllers = "ACME.Express.Controllers";
 
 @injectable()
 export class Controllers {
-  public constructor(
-    @inject(diAcmeController) protected acmeController: AcmeController,
-  ) { }
 
-  public async getDirectory(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  protected acmeController = container.resolve<AcmeController>(diAcmeController);
+
+  public async getDirectory(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.getDirectory(request);
     this.createHttpResponseMessage(result, res);
   }
-  public async getNonce(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async getNonce(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.getNonce(request);
     this.createHttpResponseMessage(result, res);
   }
 
   //#region Account
-  public async newAccount(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async newAccount(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.newAccount(request);
     this.createHttpResponseMessage(result, res);
   }
-  public async postAccount(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async postAccount(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.postAccount(request);
     this.createHttpResponseMessage(result, res);
   }
-  public async keyChange(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async keyChange(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.keyChange(request);
     this.createHttpResponseMessage(result, res);
@@ -43,47 +41,47 @@ export class Controllers {
   //#endregion
 
   //#region Order
-  public async createOrder(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async createOrder(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.createOrder(request);
     this.createHttpResponseMessage(result, res);
   }
-  public async postOrder(req: Request<ParamsDictionary>, res: Response, orderId: Key): Promise<void> {
+  public async postOrder(req: Request, res: Response, orderId: Key): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.postOrder(request, orderId);
     this.createHttpResponseMessage(result, res);
   }
-  public async postOrders(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async postOrders(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.postOrders(request);
     this.createHttpResponseMessage(result, res);
   }
-  public async finalizeOrder(req: Request<ParamsDictionary>, res: Response, orderId: Key): Promise<void> {
+  public async finalizeOrder(req: Request, res: Response, orderId: Key): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.finalizeOrder(request, orderId);
     this.createHttpResponseMessage(result, res);
   }
   //#endregion
 
-  public async postChallenge(req: Request<ParamsDictionary>, res: Response, challengeId: Key): Promise<void> {
+  public async postChallenge(req: Request, res: Response, challengeId: Key): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.postChallenge(request, challengeId);
     this.createHttpResponseMessage(result, res);
   }
 
-  public async postAuthorization(req: Request<ParamsDictionary>, res: Response, authId: Key): Promise<void> {
+  public async postAuthorization(req: Request, res: Response, authId: Key): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.postAuthorization(request, authId);
     this.createHttpResponseMessage(result, res);
   }
 
   //#region Certificate
-  public async getCertificate(req: Request<ParamsDictionary>, res: Response, thumbprint: string): Promise<void> {
+  public async getCertificate(req: Request, res: Response, thumbprint: string): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.getCertificate(request, thumbprint);
     this.createHttpResponseMessage(result, res);
   }
-  public async revokeCertificate(req: Request<ParamsDictionary>, res: Response): Promise<void> {
+  public async revokeCertificate(req: Request, res: Response): Promise<void> {
     const request = this.getAcmeRequest(req);
     const result = await this.acmeController.revokeCertificate(request);
     this.createHttpResponseMessage(result, res);
@@ -142,7 +140,7 @@ export class Controllers {
    * Parse request to AcmeRequest
    * @param req express request
    */
-  protected getAcmeRequest(req: Request<ParamsDictionary>): AcmeRequest {
+  protected getAcmeRequest(req: Request): AcmeRequest {
     const result = new AcmeRequest();
 
     // parse method
@@ -158,7 +156,7 @@ export class Controllers {
         break;
 
       default:
-        throw new MalformedError();
+        throw new MalformedError("Method is not supported");
     }
 
     // parse query
