@@ -1,17 +1,10 @@
 import { IDirectoryService } from "./types";
 import { Directory } from "@peculiar/acme-protocol";
-import { BaseService, IServerOptions, diServerOptions } from "./base";
-import { injectable, inject } from "tsyringe";
-import { diLogger, ILogger } from "@peculiar/acme-core";
+import { BaseService } from "./base";
+import { injectable } from "tsyringe";
 
 @injectable()
 export class DirectoryService extends BaseService implements IDirectoryService {
-
-  public constructor(
-    @inject(diLogger) logger: ILogger,
-    @inject(diServerOptions) options: IServerOptions) {
-    super(options, logger);
-  }
 
   public async getDirectory() {
     const url = this.options.baseAddress;
@@ -28,12 +21,12 @@ export class DirectoryService extends BaseService implements IDirectoryService {
       directory.meta = this.options.meta;
     }
 
-    this.onGetDirectory(directory);
+    await this.onGetDirectory(directory);
 
     return directory;
   }
 
-  protected onGetDirectory(directory: Directory): void {
+  protected async onGetDirectory(directory: Directory): Promise<void> {
     directory;
   }
 
