@@ -1,35 +1,30 @@
 import { IExternalAccountService } from "./types";
 import { Key, IExternalAccount, IExternalAccountRepository, diExternalAccount, diExternalAccountRepository } from "@peculiar/acme-data";
-import { cryptoProvider, MalformedError, diLogger, ILogger } from "@peculiar/acme-core";
-import { BaseService, diServerOptions, IServerOptions } from "./base";
-import { inject, container, injectable } from "tsyringe";
+import { MalformedError } from "@peculiar/acme-core";
+import { BaseService } from "./base";
+import { container, injectable } from "tsyringe";
 import { Convert } from "pvtsutils";
 import { JsonWebSignature } from "@peculiar/jose";
+import { cryptoProvider } from "@peculiar/x509";
 
 @injectable()
 export class ExternalAccountService extends BaseService implements IExternalAccountService {
 
-  public constructor(
-    @inject(diExternalAccountRepository)
-    protected externalAccountRepository: IExternalAccountRepository,
-    @inject(diLogger) logger: ILogger,
-    @inject(diServerOptions) options: IServerOptions) {
-    super(options, logger);
-  }
+protected externalAccountRepository = container.resolve<IExternalAccountRepository>(diExternalAccountRepository);
 
   public async create(account: any) {
     const macKey = cryptoProvider.get().getRandomValues(new Uint8Array(256 >> 3)); // TODO use service options for HMAC key length
 
     let externalAccount = container.resolve<IExternalAccount>(diExternalAccount);
-    this.onCreate(externalAccount, account, macKey);
+    await this.onCreate(externalAccount, account, macKey);
     externalAccount = await this.externalAccountRepository.add(externalAccount);
 
-    this.logger.info(`External account ${account.id} created`);
+    this.logger.info(`External account ${externalAccount.id} created`);
 
     return externalAccount;
   }
 
-  protected onCreate(externalAccount: IExternalAccount, account: any, key: BufferSource) {
+  protected async onCreate(externalAccount: IExternalAccount, account: any, key: BufferSource) {
     externalAccount.key = Convert.ToBase64Url(key);
     externalAccount.account = account;
     externalAccount.status = "pending";
@@ -66,9 +61,10 @@ export class ExternalAccountService extends BaseService implements IExternalAcco
     // if (!eabPayload.Equals(accountKey)) {
     //   throw new MalformedError("Signed content in externalAccountBinding doesn't match to requirement"); // TODO check rfc error
     // }
-
-    const externalAccount = await this.getById(header.kid!);
-    if (externalAccount.status != "pending") {
+    const matches = /([^/]+)$/.exec(header.kid!);
+    // TODO Add check on null
+    const externalAccount = await this.getById(matches![1]);
+    if (externalAccount.status !== "pending") {
       throw new MalformedError("External account has wrong status"); // TODO check rfc error
     }
 
